Check that /etc/crowbar/provisioner.json contains what we need
We will update the provisioner proposal with the Cloud 3 URLs for repos
if it had Cloud 2.0 URLs. To do that, we first need to check that
/etc/crowbar/provisioner.json has the URLs we need (if the provisioner
proposal is using any URL).

We also check that addition URL in /etc/crowbar/provisioner.json are
really wanted because they will be added to the provisioner proposal.