Also stop epmd after we stopped rabbitmq
There's a sleep because the unregistration of the rabbitmq name is not
effective immediately after we stopped the service.

We use rcepmd when available (not in Cloud 4, though), and "epmd -kill"
otherwise. Note that -kill will only kill it if it's not used anymore,
so it's really what we want. In the worst case, we keep the old version
running.