Manually kill neutron-ns-metadata-proxy
If set up in a HA environment, a neutron-ns-metadata-proxy process will persist
when we shut down neutron, as is needed during the upgrade process from Cloud 4
to Cloud 5. This will break the upgrade. Upstream iss working on the handling
of external processes, but in the meantime we need a workaround.

Get nodes with the neutron-l3 role and manually kill the neutron-ns-metadata-proxy
process during the upgrade.