Do not stop group resources when stopping HA resources
This will create an issue when re-applying the proposals later on, as
resources will be told to start, but this will be masked by the group;
so they won't start and any check assuming they do start will fail.

Thanks to Adam Spiers <aspiers@suse.com> for finding the bug.